% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add.edges}
\alias{add.edges}
\title{Add Edges To A Graph}
\usage{
add.edges(
  g,
  new.edges,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{new.edges}{New edges to be added (dataframe with at least two columns: 'from' and 'to').}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function adds edges to a previously learned graph.
}
\examples{
g <- add.edges(g, new.edges)
}
\keyword{edges}
\keyword{genes}
\keyword{graph}
