% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add.genes}
\alias{add.genes}
\title{Add Genes To A Graph}
\usage{
add.genes(
  g,
  new.genes,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{new.genes}{New genes to be added (vector).}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function adds genes to a previously learned graph.
}
\examples{
g <- add.genes(g, c('Cd74', 'Ldhc', 'Tlx3'))
}
\keyword{genes}
\keyword{graph}
