% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.bn}
\alias{as.bn}
\title{Convert Graph To bn Format}
\usage{
as.bn(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to bn format (bnlearn).
}
\examples{
g <- as.bn(obj)
}
