% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.bnlearn}
\alias{as.bnlearn}
\title{Convert Graph To bnlearn Format}
\usage{
as.bnlearn(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to bnlearn format (bnlearn).
}
\examples{
g <- as.bnlearn(obj)
}
