% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.igraph}
\alias{as.igraph}
\title{Convert Graph To igraph Format}
\usage{
as.igraph(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to igraph format.
}
\examples{
g <- as.igraph(obj)
}
