% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as_adjacency}
\alias{as_adjacency}
\title{Convert Graph To Adjacency Matrix}
\usage{
as_adjacency(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to adjacency matrix format.
}
\examples{
mtx <- as_adjacency(obj)
}
