% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{as_edges}
\alias{as_edges}
\title{Convert Graph To Edge List}
\usage{
as_edges(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to edge list format.
}
\examples{
mtx <- as_edges(obj)
}
