% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert Graph To igraph Format}
\usage{
as_igraph(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert your graph to igraph format.
}
\examples{
g <- as_igraph(obj)
}
