% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{averaged.graph}
\alias{averaged.graph}
\title{Calculate The Averaged Graph}
\usage{
averaged.graph(graphs, names, threshold = 0.5, to = "igraph")
}
\arguments{
\item{graphs}{List of graphs (with the same genes and in the same order).}

\item{names}{Vector with gene names (in order).}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged graph (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn') (optional).}
}
\description{
This function allows you to calculate the averaged graph from a list of graphs.
}
\examples{
graph <- averaged.graph(df)
}
\keyword{average}
\keyword{graph}
