% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.arges}
\alias{boot.arges}
\title{Adaptively Restricted Greedy Equivalence Search Algorithm (ARGES) With Bootstrapping}
\usage{
boot.arges(
  df,
  whitelist = NULL,
  blacklist = NULL,
  indep.test = pcalg::gaussCItest,
  alpha = 0.01,
  max.sx = Inf,
  adaptive = c("none", "vstructures", "triples"),
  maxDegree = integer(0),
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{indep.test}{Conditional independence test to be used (pcalg implementation). Default: pcalg::gaussCItest}

\item{alpha}{Target nominal type I error rate. Default: 0.01}

\item{max.sx}{Maximum allowed size of the conditioning sets.}

\item{adaptive}{Whether constraints should be adapted to newly detected v-structures or unshielded triples: 'none', 'vstructures', or 'triples'. Default: 'none'}

\item{maxDegree}{Parameter used to limit the vertex degree of the estimated graph. Default: integer(0)}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{to}{Output format (('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn')) (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}
}
\description{
This function allows you to learn a directed graph from a dataset using the Adaptively Restricted Greedy Equivalence Search (ARGES) algorithm.
}
\examples{
graph <- boot.arges(df)
}
\keyword{graph}
\keyword{learning}
