% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.fci}
\alias{boot.fci}
\title{Fast Causal Inference Algorithm (FCI) With Bootstrapping}
\usage{
boot.fci(
  df,
  whitelist = NULL,
  blacklist = NULL,
  indep.test = pcalg::gaussCItest,
  alpha = 0.01,
  max.sx = Inf,
  pdsep.max = Inf,
  conservative = FALSE,
  maj.rule = FALSE,
  version = c("fci", "rfci", "fci.plus"),
  type = c("normal", "anytime", "adaptive"),
  rules = rep(TRUE, 10),
  doPdsep = TRUE,
  biCC = FALSE,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4,
  seed = NULL
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{indep.test}{Conditional independence test to be used (pcalg implementation). Default: pcalg::gaussCItest}

\item{alpha}{Target nominal type I error rate. Default: 0.01}

\item{max.sx}{Maximum allowed size of the conditioning sets.}

\item{pdsep.max}{Maximum size of Possible-D-SEP for which subsets are considered as conditioning sets in the conditional independence tests.}

\item{conservative}{Whether or not the conservative PC is used (pcalg implementation). Default: FALSE}

\item{maj.rule}{Whether or not the triples shall be checked for ambiguity using a majority rule idea, which is less strict than the conservative PC algorithm (pcalg implementation). Default: FALSE}

\item{version}{Version of FCI algorithm to be used: 'fci', 'rfci', or 'fci.plus'. Default: 'fci'}

\item{type}{Type of FCI algorithm to be used: 'normal', 'anytime', or 'adaptive'. Default: 'normal'}

\item{rules}{Logical vector of length 10 indicating which rules should be used when directing edges. Default: rep(TRUE,10)}

\item{doPdsep}{If FALSE, Possible-D-SEP is not computed, so that the algorithm simplifies to the Modified PC algorithm of Spirtes, Glymour and Scheines (2000, p.84). Default: TRUE}

\item{biCC}{If TRUE, only nodes on paths between nodes x and y are considered to be in Possible-D-SEP(x) when testing independence between x and y. Default: TRUE}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Fast Causal Inference algorithm (stable version).
}
\examples{
graph <- boot.fci(df)
}
\keyword{graph}
\keyword{learning}
