% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.genie3}
\alias{boot.genie3}
\title{GEne Network Inference with Ensemble of trees (GENIE3) with Bootstrapping}
\usage{
boot.genie3(
  df,
  tree.method = c("rf", "et"),
  K = "sqrt",
  n.trees = 1000,
  min.weight = 0.1,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4,
  seed = sample(1:10^10, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{tree.method}{Random Forest ('rf') or Extra-Trees ('et'). Default: 'rf'}

\item{K}{Number of candidate regulators that are randomly selected at each tree node for the best split determination. Default: 'sqrt' (square root of the number of genes)}

\item{n.trees}{Number of trees that are grown per ensemble. Default: 1000}

\item{min.weight}{Minimum absolute value considered in the adjacency matrix. Lower values will be replaced by zero. Default: 0.1}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the GENIE3 algorithm.
}
\examples{
obj <- boot.genie3(df)
avg.g <- obj$average
g.rep <- obj$replicates
}
\keyword{graph}
\keyword{learning}
