% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.hc}
\alias{boot.hc}
\title{Hill-Climbing Algorithm (HC) With Bootstrapping}
\usage{
boot.hc(
  df,
  start = NULL,
  whitelist = NULL,
  blacklist = NULL,
  score = scores,
  restart = 0,
  perturb = 1,
  max.iter = Inf,
  maxp = Inf,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = parallel::detectCores(),
  seed = sample(1:10^6, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{start}{Preseeded directed acyclic graph used to initialize the algorithm (optional).}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{score}{Score to be used: 'pred-loglik-g', 'loglik-g', 'aic-g', 'bic-g', or 'bge'. Default: 'pred-loglik-g'}

\item{restart}{Number of random restarts. Default: 0}

\item{perturb}{Number of attempts to randomly insert/remove/reverse an arc on every random restart. Default: 1}

\item{max.iter}{Maximum number of iterations. Default: Inf}

\item{maxp}{Maximum number of parents for a node. Default: Inf}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: parallel::detectCores()}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Hill-Climbing algorithm.
}
\examples{
obj <- boot.hc(df)
avg.g <- obj$average
g.rep <- obj$replicates
}
\keyword{graph}
\keyword{learning}
