% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.lingam}
\alias{boot.lingam}
\title{LINGAM Algorithm With Bootstrapping}
\usage{
boot.lingam(df, R = 200, m = NULL, threshold = 0.5, to = "igraph", cluster = 4)
}
\arguments{
\item{df}{Dataset.}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}
}
\description{
This function allows you to learn a directed graph from a dataset using the LINGAM algorithm.
}
\examples{
graph <- boot.lingam(df)
}
\keyword{graph}
\keyword{learning}
