% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.nodag}
\alias{boot.nodag}
\title{NODAG Algorithm With Bootstrapping}
\usage{
boot.nodag(
  lib.path,
  df,
  lambda = 0.5,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4
)
}
\arguments{
\item{lib.path}{Path of 'nodag.so' file.}

\item{df}{Dataset.}

\item{lambda}{Lambda regularization parameter. Default: 0.5}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{to}{Output format (('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn')) (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}
}
\description{
This function allows you to learn a directed graph from a dataset using the NODAG algorithm.
}
\examples{
graph <- boot.nodag('nodag.so', df)
}
\keyword{graph}
\keyword{learning}
