% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.notears}
\alias{boot.notears}
\title{Linear NO-TEARS Algorithm (Reimplemented) With Bootstrapping}
\usage{
boot.notears(
  df,
  lambda1 = 0.1,
  loss.type = c("l2", "logistic", "poisson"),
  max.iter = 100,
  h.tol = 1e-08,
  rho.max = 1e+16,
  w.threshold = 0.3,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4,
  seed = NULL
)
}
\arguments{
\item{df}{Dataset.}

\item{lambda1}{L1 regularization parameter. Default: 0.1}

\item{loss.type}{Type of loss function to be used: 'l2', 'logistic', or 'poisson'. Default: 'l2'}

\item{max.iter}{Maximum number of dual ascent steps. Default: 100}

\item{h.tol}{Minimum absolute value of h. Default: 1e-8}

\item{rho.max}{Maximum value of rho. Default: 1e+16}

\item{w.threshold}{Threshold of absolute value of weight. Default: 0.3}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Linear NO-TEARS algorithm.
}
\examples{
obj <- boot.notears(df)
avg.g <- obj$average
g.rep <- obj$replicates
}
\keyword{graph}
\keyword{learning}
