% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.rsmax2}
\alias{boot.rsmax2}
\title{General 2-Phase Restricted Maximization Algorithm (rsmax2) With Bootstrapping}
\usage{
boot.rsmax2(
  df,
  whitelist = NULL,
  blacklist = NULL,
  restrict = c("pc.stable", "gs", "iamb", "fast.iamb", "inter.iamb", "iamb.fdr",
    "mmpc", "si.hiton.pc", "hpc"),
  maximize = c("hc", "tabu"),
  restrict.args = list(),
  maximize.args = list(),
  version = c("rsmax2", "mmhc", "h2pc"),
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = "igraph",
  cluster = 4
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{restrict}{Constraint-based or local search algorithm to be used in the restrict phase: 'pc.stable', 'gs', 'iamb', 'fast.iamb', 'inter.iamb', 'iamb.fdr', 'mmpc', 'si.hiton.pc', or 'hpc'. Default: 'pc.stable'}

\item{maximize}{Score-based algorithm to be used in the maximize phase: 'hc' or 'tabu'. Default: 'hc'}

\item{restrict.args}{List of arguments to be passed to the algorithm specified by restrict.}

\item{maximize.args}{List of arguments to be passed to the algorithm specified by maximize.}

\item{version}{Algorithm version: 'rsmax2','mmhc', or 'h2pc'. Default: 'rsmax2'}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}
}
\description{
This function allows you to learn a directed graph from a dataset using the General 2-Phase Restricted Maximization algorithm.
}
\examples{
graph <- boot.rsmax2(df)
}
\keyword{graph}
\keyword{learning}
