% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{chowliu}
\alias{chowliu}
\title{Chow-Liu Algorithm}
\usage{
chowliu(
  df,
  whitelist = NULL,
  blacklist = NULL,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = sample(1:10^6, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph.}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph.}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a undirected graph from a dataset using the Chow-Liu algorithm.
}
\examples{
g <- chowliu(df)
}
\keyword{graph}
\keyword{learning}
