% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{community.plot}
\alias{community.plot}
\title{Community Plotting}
\usage{
community.plot(
  x,
  algorithm = c("louvain", "edge.betweenness", "fast.greedy", "label.prop",
    "leading.eigen", "optimal", "spinglass", "walktrap"),
  from = NULL,
  layout = c("grid", "star", "circle", "tree", "nicely"),
  interactive = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{algorithm}{Algorithm for finding communities: 'louvain', 'edge.betweenness', 'fast.greedy', 'label.prop', 'leading.eigen', 'optimal', 'spinglass', or 'walktrap'. Default: 'louvain'}

\item{from}{Input format (optional).}

\item{layout}{igraph plot layout (optional): 'grid', 'star', 'circle', 'tree', or 'nicely'. It will be ignored if interactive=TRUE. Default: 'grid'}

\item{interactive}{Interactive plot (optional). Default: FALSE}
}
\description{
This function allows you to plot a graph with each community in a different color.
}
\examples{
community.plot(obj, interactive=FALSE)
community.plot(obj, interactive=TRUE)
}
\keyword{community}
\keyword{graph}
\keyword{plot}
