% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert.format}
\alias{convert.format}
\title{Graph Format Conversion}
\usage{
convert.format(x, to, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bnlearn').}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert graph between different formats: adjacency matrix, list of edges, igraph and bnlearn (bnlearn).
}
\examples{
graph <- convert.format(mtx, 'igraph')
}
\keyword{adjacency}
\keyword{edges}
\keyword{format}
\keyword{graph}
