% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{convert_format}
\alias{convert_format}
\title{Graph Format Conversion}
\usage{
convert_format(x, to, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn').}

\item{from}{Input format (optional).}
}
\description{
This function allows you to convert graph between different formats: adjacency matrix, list of edges, igraph and bn (bnlearn).
}
\examples{
graph <- convert_format(mtx, 'igraph')
}
\keyword{adjacency}
\keyword{edges}
\keyword{format}
\keyword{graph}
