% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset.split}
\alias{dataset.split}
\title{Split A Dataset Into Training And Test Datasets}
\usage{
dataset.split(df, m = NULL)
}
\arguments{
\item{df}{Dataset.}

\item{m}{Size of training dataset (optional). Default: nrow(df)/2}
}
\description{
This function allows you to split a dataset into training and test datasets.
}
\examples{
splitted.df <- dataset.split(df)
splitted.df$train
splitted.df$test
}
\keyword{dataframe}
\keyword{test}
\keyword{training}
