% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{delete.edges}
\alias{delete.edges}
\title{Delete Edges To A Graph}
\usage{
delete.edges(
  g,
  rm.edges,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{rm.edges}{New edges to be deleted (dataframe with at least two columns: 'from' and 'to').}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function deletes edges to a previously learned graph.
}
\examples{
g <- delete.edges(g, new.edges)
}
\keyword{edges}
\keyword{genes}
\keyword{graph}
