% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{delete.genes}
\alias{delete.genes}
\title{Delete Genes From A Graph}
\usage{
delete.genes(
  g,
  rm.genes,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{rm.genes}{Genes to be deleted (vector).}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function deletes genes from a previously learned graph.
}
\examples{
g <- delete.genes(g, c('Cd74', 'Ldhc', 'Tlx3'))
}
\keyword{genes}
\keyword{graph}
