% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{delete.isolated}
\alias{delete.isolated}
\title{Delete Isolated Nodes}
\usage{
delete.isolated(
  g,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function allows you to delete isolated genes (nodes) in a graph.
}
\examples{
g <- delete.isolated(g, to='igraph')
}
\keyword{genes}
\keyword{graph}
\keyword{isolated}
