% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{detect.format}
\alias{detect.format}
\title{Graph Format Detection}
\usage{
detect.format(x)
}
\arguments{
\item{x}{Graph object.}
}
\description{
This function allows you to detect graph format.
}
\examples{
detect.format(obj)
}
\keyword{format}
\keyword{graph}
