% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{detect_format}
\alias{detect_format}
\title{Graph Format Detection}
\usage{
detect_format(x)
}
\arguments{
\item{x}{Graph object.}
}
\description{
This function allows you to detect graph format.
}
\examples{
detect_format(obj)
}
\keyword{format}
\keyword{graph}
