% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download.geneset}
\alias{download.geneset}
\title{Download A Geneset Via RESTful API}
\usage{
download.geneset(code, host = HOST)
}
\arguments{
\item{code}{Download Code (indicated by list.genesets() output).}
}
\description{
This function allows you to download a geneset available via RESTful API.
}
\examples{
df <- download.geneset(1)
}
\keyword{api}
\keyword{genes}
