% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download.graph}
\alias{download.graph}
\title{Download A Graph Via RESTful API}
\usage{
download.graph(
  code,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  host = HOST
)
}
\arguments{
\item{code}{Download Code (indicated by list.graphs() output).}
}
\description{
This function allows you to download a graph available via RESTful API.
}
\examples{
g <- download.graph(1)
}
\keyword{api}
\keyword{graphs}
