% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_dataset}
\alias{download_dataset}
\title{Download A Dataset Via RESTful API}
\usage{
download_dataset(code, log = FALSE, host = HOST)
}
\arguments{
\item{code}{Download Code (indicated by list_datasets() output).}
}
\description{
This function allows you to download a dataset available via RESTful API.
}
\examples{
df <- download_dataset(1)
}
\keyword{api}
\keyword{datasets}
