% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_geneset}
\alias{download_geneset}
\title{Download A Geneset Via RESTful API}
\usage{
download_geneset(code, host = HOST)
}
\arguments{
\item{code}{Download Code (indicated by list_genesets() output).}
}
\description{
This function allows you to download a geneset available via RESTful API.
}
\examples{
df <- download_geneset(1)
}
\keyword{api}
\keyword{genes}
