% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{drop.all.zeros}
\alias{drop.all.zeros}
\title{Drop rows and/or columns with all zeros}
\usage{
drop.all.zeros(mtx, rows = TRUE, columns = TRUE, square.matrix = "none")
}
\arguments{
\item{mtx}{Matrix or dataframe.}

\item{rows}{Drop rows with all zeros.}

\item{columns}{Drop columns with all zeros.}

\item{square.matrix}{Special mode for square matrices ('any' or 'all'). any: if the column or row is full of zeros, delete both; all: if the column or row is not full of zeros, keep both}
}
\description{
This function allows you to drop rows and/or columns with all zeros.
}
\examples{
mtx <- drop.all.zeros(mtx)
mtx <- drop.all.zeros(mtx, square.matrix='any')
mtx <- drop.all.zeros(mtx, square.matrix='all')
}
\keyword{dataframe}
\keyword{matrix}
\keyword{zeros}
