% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{export.dataset}
\alias{export.dataset}
\title{Export A Dataset To A File}
\usage{
export.dataset(
  df,
  path,
  sep = "\\t",
  header = TRUE,
  index = FALSE,
  transpose = FALSE
)
}
\arguments{
\item{df}{Dataset to export.}

\item{path}{Output file path.}

\item{sep}{Separator / delimiter character (optional). Default: TAB}

\item{header}{Whether or not to include a header (optional). Default: TRUE}

\item{index}{Whether or not to include an index column (optional). Default: FALSE}

\item{transpose}{Whether or not to transpose the data matrix (optional). IMPORTANT NOTE: gnlearn works with cell x gene matrices. Default: FALSE}
}
\description{
This function allows you to export a dataset to a file.
}
\examples{
export.dataset(df, 'mydataset.tsv')
}
\keyword{datasets}
