% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{export.geneset}
\alias{export.geneset}
\title{Export A Geneset To A File}
\usage{
export.geneset(df, path, sep = "\\t", header = TRUE, index = FALSE)
}
\arguments{
\item{df}{Geneset to export.}

\item{path}{Output file path.}

\item{sep}{Separator / delimiter character (optional). Default: TAB}

\item{header}{Whether or not to include a header (optional). Default: TRUE}

\item{index}{Whether or not to include an index column (optional). Default: FALSE}
}
\description{
This function allows you to export a geneset to a file.
}
\examples{
export.geneset(df, 'mygeneset.tsv')
}
\keyword{genesets}
