% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{export.graph}
\alias{export.graph}
\title{Export A Graph To A File}
\usage{
export.graph(
  g,
  path,
  sep = "\\t",
  header = TRUE,
  index = TRUE,
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn"),
  to = c("adjacency", "edges")
)
}
\arguments{
\item{g}{Graph to export.}

\item{path}{Output file path.}

\item{sep}{Separator / delimiter character (optional). Default: TAB}

\item{header}{Whether or not to include a header (optional). Default: TRUE}

\item{index}{Whether or not to include an index column (optional). Default: FALSE}

\item{from}{Output format (optional): 'adjacency' or 'edges'. Default: 'adjacency'}
}
\description{
This function allows you to graph a geneset to a file.
}
\examples{
export.graph(g, 'mygraph.tsv')
}
\keyword{graphs}
