% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{feature.degree}
\alias{feature.degree}
\title{Feature Degree}
\usage{
feature.degree(
  x,
  genes,
  features = NULL,
  vertices = NULL,
  in.degree = TRUE,
  out.degree = TRUE,
  loops = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{genes}{Geneset with features.}

\item{features}{Features you want to analyze. Default: all boolean features.}

\item{vertices}{Vertices you want to analyze. Default: all vertices.}

\item{in.degree}{Whether or not to analyze in-degree (optional). Default: TRUE.}

\item{out.degree}{Whether or not to analyze out-degree (optional). Default: TRUE.}

\item{loops}{Whether or not to consider loops (optional). Default: FALSE.}

\item{normalized}{Whether or not to normalize degrees (optional). Default: FALSE.}
}
\description{
This function allows you to calculate the in-degree and out-degree of genes that have a certain feature (e.g. transcription factors or tumor suppressor genes).
}
\examples{
mtx <- feature.degree(x, genes)
mtx <- feature.degree(x, genes, features=c('tf', 'target', 'tumor.suppressor'))
mtx <- feature.degree(x, genes, features='tumor.suppressor', in.degree=TRUE, out.degree=FALSE)
mtx <- feature.degree(x, genes, features='essential', vertices=c('ADRB1', 'HSF2'), normalized=TRUE)
}
\keyword{degree}
\keyword{feature}
\keyword{graph}
\keyword{vertex}
