% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{feature.plot}
\alias{feature.plot}
\title{Feature Graph Plotting}
\usage{
feature.plot(
  x,
  genes,
  feature,
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn"),
  layout = c("circle", "star", "grid", "sphere", "nicely"),
  interactive = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{genes}{Geneset with features.}

\item{from}{Input format (optional).}

\item{layout}{igraph plot layout (optional): 'grid', 'star', 'circle', 'sphere', or 'nicely'. Default: 'circle'}

\item{interactive}{Interactive plot (optional). Default: FALSE}

\item{features}{Feature whose graph you want to plot.}
}
\description{
This function allows you to plot the graph of a feature (e.g. transcription factors or tumor suppressors).
}
\examples{
feature.plot(obj, genes, 'tf', interactive=FALSE)
feature.plot(obj, genes, 'tumor.suppressor', interactive=TRUE)
}
\keyword{feature}
\keyword{graph}
\keyword{plot}
