% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{feature_degree}
\alias{feature_degree}
\title{Feature Degree}
\usage{
feature_degree(
  x,
  genes,
  features = NULL,
  vertices = NULL,
  in.degree = TRUE,
  out.degree = TRUE,
  loops = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{genes}{Geneset with features.}

\item{features}{Features you want to analyze. Default: all boolean features.}
}
\description{
This function allows you to calculate the in-degree and out-degree of genes that have a certain feature (e.g. transcription factors or tumor suppressor genes).
}
\examples{
mtx <- feature_degree(x, genes)
mtx <- feature_degree(x, genes, features=c('tf', 'target', 'tumor.suppressor'))
mtx <- feature_degree(x, genes, features='tumor.suppressor', in.degree=TRUE, out.degree=FALSE)
mtx <- feature_degree(x, genes, features='essential', vertices=c('ADRB1', 'HSF2'), normalized=TRUE)
}
\keyword{Default:}
\keyword{Vertices}
\keyword{all}
\keyword{analyze.}
\keyword{to}
\keyword{vertices}
\keyword{vertices.}
\keyword{want}
\keyword{you}
