% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{filter.dataset}
\alias{filter.dataset}
\title{Select (automatically) the most appropriate gene columns and cell rows of your dataset.}
\usage{
filter.dataset(
  df,
  filter.first = c("cells", "genes"),
  genes = NULL,
  selected.genes = NULL,
  features = NULL,
  max.genes = 100,
  max.cells = 2000,
  genes.criteria = c("non.zeros", "mean", "variance", "random"),
  cells.criteria = c("non.zeros", "mean", "variance", "random"),
  min.non.zeros = 2,
  cor = TRUE,
  cor.method = c("spearman", "kendall", "pearson"),
  cor.threshold = 0.5
)
}
\arguments{
\item{df}{Dataset.}

\item{filter.first}{Whether to filter the cells or the genes first. Default: cells}

\item{genes}{Geneset with features (optional).}

\item{selected.genes}{User-selected genes (optional).}

\item{features}{Features you want to select (optional).}

\item{max.genes}{Maximum number of selected genes (optional).}

\item{max.cells}{Maximum number of selected cells (optional).}

\item{genes.criteria}{Criteria to be applied when it is necessary to select a maximum number of genes from the previously filtered ones. Default: 'non.zeros'}

\item{cells.criteria}{Criteria to be applied to select cells. Default: 'non.zeros'}

\item{min.non.zeros}{Minimum number of non-zero values per gene (optional).}

\item{cor}{Using correlation to exclude unconnected genes (optional).}

\item{cor.method}{Correlation method: 'spearman', 'kendall', or 'pearson'. Default: 'spearman'}

\item{cor.threshold}{Correlation threshold. Default: 0.1}
}
\description{
This function allows you to (automatically) select the most appropriate gene columns and cell rows of your dataset.
}
\examples{
df <- filter.dataset(df, max.genes=100, max.cells=2000)
}
\keyword{genes}
\keyword{select}
