% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{gclm}
\alias{gclm}
\title{Graphical Continuous Lyapunov Models (GCLM)}
\usage{
gclm(
  df,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = sample(1:10^10, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the GCLM algorithm.
}
\examples{
g <- gclm(df)
}
\keyword{graph}
\keyword{learning}
