% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gene.correlation}
\alias{gene.correlation}
\title{Plot expression correlation of genes in your dataset.}
\usage{
gene.correlation(
  df,
  selected.genes = NULL,
  method = c("spearman", "kendall", "pearson")
)
}
\arguments{
\item{df}{Dataset.}

\item{selected.genes}{User-selected genes (optional).}

\item{method}{Correlation method: 'spearman', 'kendall', or 'pearson'. Default: 'spearman'}
}
\description{
This function allows you to plot the expression correlation of genes in your dataset.
}
\examples{
gene.correlation(df)
}
\keyword{correlation}
\keyword{expression}
\keyword{genes}
