% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gene.degree}
\alias{gene.degree}
\title{Degree Per Gene}
\usage{
gene.degree(
  x,
  vertices = NULL,
  in.degree = TRUE,
  out.degree = TRUE,
  loops = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{vertices}{Vertices you want to analyze. Default: all vertices.}

\item{in.degree}{Whether or not to analyze in-degree (optional). Default: TRUE.}

\item{out.degree}{Whether or not to analyze out-degree (optional). Default: TRUE.}

\item{loops}{Whether or not to consider loops (optional). Default: FALSE.}

\item{normalized}{Whether or not to normalize degrees (optional). Default: FALSE.}
}
\description{
This function allows you to calculate the in-degree and out-degree of each (selected) gene.
}
\examples{
mtx <- gene.degree(x)
}
\keyword{degree}
\keyword{graph}
\keyword{vertex}
