% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gene.histogram}
\alias{gene.histogram}
\title{Plot expression histograms of genes in your dataset.}
\usage{
gene.histogram(df, selected.genes = NULL, n.cols = 3)
}
\arguments{
\item{df}{Dataset.}

\item{selected.genes}{User-selected genes (optional).}

\item{n.cols}{Number of column in the plot (optional). Default: 3}
}
\description{
This function allows you to plot the expression histograms of genes in your dataset.
}
\examples{
gene.histogram(df)
}
\keyword{expression}
\keyword{genes}
\keyword{plot}
