% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gene.select}
\alias{gene.select}
\title{Select (automatically) the most appropriate gene columns of your dataset.}
\usage{
gene.select(
  df,
  genes = NULL,
  selected.genes = NULL,
  features = NULL,
  max.genes = 100,
  min.non.zeros = 2,
  glasso = TRUE
)
}
\arguments{
\item{df}{Dataset.}

\item{genes}{Geneset with features (optional).}

\item{selected.genes}{User-selected genes (optional).}

\item{features}{Features you want to select (optional).}

\item{max.genes}{Maximum number of selected genes (optional).}

\item{min.non.zeros}{Minimum number of non-zero values per gene (optional).}

\item{glasso}{Using the GLASSO algorithm to exclude unconnected genes (optional).}
}
\description{
This function allows you to (automatically) select the most appropriate gene columns of your dataset.
}
\examples{
df <- gene.select(df, max.genes=100, min.non.zeros=2)
df <- gene.select(df, genes=genes, features=c('tumor.suppressor', 'breast.cancer'), glasso=TRUE, rho=0.5)
}
\keyword{genes}
\keyword{select}
