% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{ges}
\alias{ges}
\title{Greedy Equivalence Search Algorithm (GES) With Bootstrapping}
\usage{
ges(
  df,
  blacklist = NULL,
  adaptive = c("none", "vstructures", "triples"),
  maxDegree = integer(0),
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = sample(1:10^10, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{adaptive}{Whether constraints should be adapted to newly detected v-structures or unshielded triples: 'none', 'vstructures', or 'triples'. Default: 'none'}

\item{maxDegree}{Parameter used to limit the vertex degree of the estimated graph. Default: integer(0)}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Greedy Equivalence Search (GES) algorithm of Chickering (2002).
}
\examples{
g <- ges(df)
}
\keyword{graph}
\keyword{learning}
