% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{glasso}
\alias{glasso}
\title{Graphical Lasso (GLASSO)}
\usage{
glasso(
  df,
  rho = 0.1,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = sample(1:10^10, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{rho}{Non-negative regularization parameter for GLASSO.}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn an undirected graph from a dataset using the GLASSO algorithm.
}
\examples{
g <- glasso(df, rho=0.1)
}
\keyword{graph}
\keyword{learning}
