% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.marginalization}
\alias{graph.marginalization}
\title{Graph Marginalization}
\usage{
graph.marginalization(
  g,
  obs.genes,
  max.steps = Inf,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{obs.genes}{Vector of observed genes.}

\item{max.steps}{Maximum number of steps in the path. Default: Inf}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}
}
\description{
This function allows you to marginalize a graph over observed genes.
}
\examples{
g <- graph.marginalization(g, obs.genes)
}
\keyword{genes}
\keyword{graph}
\keyword{marginalization}
