% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.plot}
\alias{graph.plot}
\title{Graph Plotting}
\usage{
graph.plot(
  x,
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn"),
  layout = c("grid", "star", "circle", "sphere", "nicely"),
  interactive = FALSE,
  isolated.genes = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}

\item{layout}{igraph plot layout (optional): 'grid', 'star', 'circle', 'sphere', or 'nicely'. Default: 'grid'}

\item{interactive}{Interactive plot (optional). Default: FALSE}

\item{isolated.genes}{Whether or not to include isolated nodes in the plot (optional). Default: FALSE}
}
\description{
This function allows you to plot a graph.
}
\examples{
graph.plot(obj)
graph.plot(obj, isolated.genes=TRUE)
graph.plot(obj, interactive=TRUE)
}
\keyword{graph}
\keyword{plot}
