% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ground.truth}
\alias{ground.truth}
\title{Create a Ground Truth Graph}
\usage{
ground.truth(x, to = c("igraph", "adjacency", "edges", "graph", "bnlearn"))
}
\arguments{
\item{x}{Download Code (must be of type 'TF-Target Interactions') or list of edges (data.frame or matrix).}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn') (optional).}
}
\description{
This function allows you to create a ground truth graph from a geneset available via RESTful API or from a local list of edges.
}
\examples{
gt <- ground.truth(2)
}
\keyword{graph}
\keyword{groundtruth}
