% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{hc}
\alias{hc}
\title{Hill-Climbing Algorithm (HC)}
\usage{
hc(
  df,
  start = NULL,
  whitelist = NULL,
  blacklist = NULL,
  score = scores,
  restart = 0,
  perturb = 1,
  max.iter = Inf,
  maxp = Inf,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = NULL
)
}
\arguments{
\item{df}{Dataset.}

\item{start}{Preseeded directed acyclic graph used to initialize the algorithm (optional).}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{score}{Score to be used: 'pred-loglik-g', 'loglik-g', 'aic-g', 'bic-g', or 'bge'. Default: 'pred-loglik-g'}

\item{restart}{Number of random restarts. Default: 0}

\item{perturb}{Number of attempts to randomly insert/remove/reverse an arc on every random restart. Default: 1}

\item{max.iter}{Maximum number of iterations. Default: Inf}

\item{maxp}{Maximum number of parents for a node. Default: Inf}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Hill-Climbing algorithm.
}
\examples{
g <- hc(df)
}
\keyword{graph}
\keyword{learning}
