% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{heatmap.adjacency}
\alias{heatmap.adjacency}
\title{Adjacency Matrix Plotting}
\usage{
heatmap.adjacency(x, from = NULL, col = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}

\item{col}{Heatmap color palette (optional).}
}
\description{
This function allows you to plot a graph as an adjacency matrix, regardless of the format (adjacency matrix, list of edges, igraph, or bn).
}
\examples{
heatmap.adjacency(obj)
}
\keyword{adjacency}
\keyword{heatmap}
\keyword{plot}
