% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{huge.graph}
\alias{huge.graph}
\title{Learn Huge Graph (With Random Gene Selection + Cells Bootstrapping)}
\usage{
huge.graph(
  df,
  algorithm = boot.pc,
  n.genes = 15,
  R = 200,
  threshold = 0.5,
  iter.R = 4,
  iter.m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4,
  ...
)
}
\arguments{
\item{df}{Dataset.}

\item{algorithm}{Algorithm to be used (any of the gnlearn 'boot.x' algorithms, such as boot.pc or boot.hc). Default: boot.pc}

\item{n.genes}{Number of random genes per iteration. Default: 15}

\item{R}{Number of iterations. Defaults: 200}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{iter.R}{Number of bootstrap replicates. Default: 200}

\item{iter.m}{Size of each bootstrap replicate. Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn').}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}

\item{...}{Other arguments for the specified algorithm.}
}
\description{
This function allows you to learn a directed graph from a high-dimensional dataset.
}
\examples{
graph <- huge.graph(df, algorithm=boot.tabu)
graph <- huge.graph(df, algorithm=boot.lingam)
graph <- huge.graph(df, algorithm=boot.iamb, n.genes=20, R=100, threshold=0.9, iter.R=10)
}
\keyword{graph}
\keyword{huge}
\keyword{learning}
