% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{import.geneset}
\alias{import.geneset}
\title{Import A Local Geneset}
\usage{
import.geneset(path, sep = "\\t", header = TRUE, index = FALSE)
}
\arguments{
\item{path}{Path to local geneset.}

\item{sep}{Separator / delimiter character (optional). Default: TAB}

\item{header}{Whether or not the file has a header (optional). Default: TRUE}

\item{index}{Whether or not the file has an index column (optional). Default: FALSE}
}
\description{
This function allows you to import a local geneset.
}
\examples{
df <- import.geneset('mygeneset.tsv')
}
\keyword{genesets}
