% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{import.graph}
\alias{import.graph}
\title{Import A Local Graph}
\usage{
import.graph(
  path,
  sep = "\\t",
  header = TRUE,
  index = TRUE,
  from = c("adjacency", "edges"),
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn")
)
}
\arguments{
\item{path}{Path to local graph.}

\item{sep}{Separator / delimiter character (optional). Default: TAB}

\item{header}{Whether or not the file has a header (optional). Default: TRUE}

\item{index}{Whether or not the file has an index column (optional). Default: TRUE}

\item{from}{Input format (optional): 'adjacency' or 'edges'. Default: 'adjacency'}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}
}
\description{
This function allows you to import a local graph.
}
\examples{
g <- import.graph('mygraph.tsv')
}
\keyword{graphs}
