% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{import_dataset}
\alias{import_dataset}
\title{Import A Local Dataset}
\usage{
import_dataset(
  path,
  log = FALSE,
  sep = "\\t",
  header = TRUE,
  index = FALSE,
  transpose = FALSE
)
}
\arguments{
\item{path}{Path to local dataset.}

\item{log}{Whether or not to apply log(x+1) (optional). Default: FALSE}

\item{sep}{Separator / delimiter character (optional). Default: '\t'}

\item{header}{Whether or not the file has a header (optional). Default: TRUE}

\item{index}{Whether or not the file has an index column (optional). Default: FALSE}

\item{transpose}{Whether or not to transpose the data matrix (optional). IMPORTANT NOTE: gnlearn works with cell x gene matrices. Default: FALSE}
}
\description{
This function allows you to import a local dataset.
}
\examples{
df <- import_dataset('mydataset.tsv')
}
\keyword{datasets}
