% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list.genesets}
\alias{list.genesets}
\title{List Genesets Available Via RESTful API}
\usage{
list.genesets(
  sp.scientific = NULL,
  sp.common = NULL,
  dataset = NULL,
  host = HOST
)
}
\arguments{
\item{sp.scientific}{Species (Scientific Name) (regex) (optional).}

\item{sp.common}{Species (Common Name) (regex) (optional).}

\item{dataset}{Dataset Name (regex) (optional).}
}
\description{
This function allows you to list genesets available via RESTful API.
}
\examples{
genesets <- list.genesets()
genesets <- list.genesets(sp.common='human')
genesets <- list.genesets(sp.common='mouse')
}
\keyword{api}
\keyword{genes}
