% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list.graphs}
\alias{list.graphs}
\title{List Graphs Available Via RESTful API}
\usage{
list.graphs(
  sp.scientific = NULL,
  sp.common = NULL,
  dataset = NULL,
  bio.layer = NULL,
  cell.identity = NULL,
  algorithm = NULL,
  min.nodes = 0,
  max.nodes = Inf,
  min.edges = 0,
  max.edges = Inf,
  host = HOST
)
}
\arguments{
\item{sp.scientific}{Species (Scientific Name) (regex) (optional).}

\item{sp.common}{Species (Common Name) (regex) (optional).}

\item{dataset}{Dataset used as input to learn the graph (Dataset Download Code) (optional).}

\item{bio.layer}{Biological Information Layer (regex) (optional).}

\item{cell.identity}{Biological Cell Identity (regex) (optional).}

\item{algorithm}{Algorithm used to learn the graph (regex) (optional).}

\item{min.nodes}{Minimum Number of Nodes (optional).}

\item{max.nodes}{Maximum Number of Nodes (optional).}

\item{min.edges}{Minimum Number of Edges (optional).}

\item{max.edges}{Maximum Number of Edges (optional).}
}
\description{
This function allows you to list graphs available via RESTful API.
}
\examples{
graphs <- list.graphs()
graphs <- list.graphs(sp.common='human')
graphs <- list.graphs(sp.common='mouse')
}
\keyword{api}
\keyword{graphs}
