% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_datasets}
\alias{list_datasets}
\title{List Datasets Available Via RESTful API}
\usage{
list_datasets(
  sp.scientific = NULL,
  sp.common = NULL,
  bio.layer = NULL,
  seq.protocol = NULL,
  cell.identity = NULL,
  min.genes = 0,
  max.genes = Inf,
  min.cells = 0,
  max.cells = Inf,
  host = HOST
)
}
\arguments{
\item{sp.scientific}{Species (Scientific Name) (regex) (optional).}

\item{sp.common}{Species (Common Name) (regex) (optional).}

\item{bio.layer}{Biological Information Layer (regex) (optional).}

\item{seq.protocol}{Sequencing Protocol (regex) (optional).}

\item{cell.identity}{Biological Cell Identity (regex) (optional).}

\item{min.genes}{Minimum Number of Genes (optional).}

\item{max.genes}{Maximum Number of Genes (optional).}

\item{min.cells}{Minimum Number of Cells (optional).}

\item{max.cells}{Maximum Number of Cells (optional).}
}
\description{
This function allows you to list datasets available via RESTful API.
}
\examples{
datasets <- list_datasets()
datasets <- list_datasets(sp.common='human')
datasets <- list_datasets(sp.common='mouse')
}
\keyword{api}
\keyword{datasets}
