% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make.edgelist}
\alias{make.edgelist}
\title{Make an edgelist from some genes to anothers.}
\usage{
make.edgelist(
  genes,
  from.genes = NULL,
  to.genes = NULL,
  from.features = NULL,
  to.features = NULL
)
}
\arguments{
\item{genes}{Geneset with features.}

\item{from.genes}{User-selected 'from' genes.}

\item{to.genes}{User-selected 'to' genes.}

\item{from.features}{The edges will go from genes with some of these features.}

\item{to.features}{The edges will go to genes with some of these features.}
}
\description{
This function allows you to make an edgelist from some genes to anothers.
}
\examples{
blacklist <- make.edgelist(genes, from.features='target', to.features='tf')
}
\keyword{blacklist}
\keyword{genes}
\keyword{whitelist}
