% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{nodag}
\alias{nodag}
\title{NODAG Algorithm}
\usage{
nodag(
  df,
  lambda = 0.5,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = NULL
)
}
\arguments{
\item{df}{Dataset.}

\item{lambda}{Lambda regularization parameter. Default: 0.5}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the NODAG algorithm.
}
\examples{
g <- nodag(df)
}
\keyword{graph}
\keyword{learning}
