% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ortholog.graph}
\alias{ortholog.graph}
\title{Ortholog Genes Graph}
\usage{
ortholog.graph(
  g,
  genes,
  column,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{genes}{Geneset with features.}

\item{column}{Name of column with ortholog genes.}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function allows you to convert graph between different formats: adjacency matrix, list of edges, igraph and bnlearn (bnlearn).
}
\examples{
g <- ortholog.graph(g, genes, column='human.ortholog')
}
\keyword{genes}
\keyword{graph}
\keyword{ortholog}
