% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.graph}
\alias{plot.graph}
\title{Graph Plotting}
\usage{
\method{plot}{graph}(x, from = NULL, layout = NULL, interactive = FALSE)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}

\item{layout}{igraph plot layout (optional). It will be ignored if interactive=TRUE.}

\item{interactive}{Interactive plot (optional). Default: FALSE}
}
\description{
This function allows you to plot a graph, regardless of the format (adjacency matrix, list of edges, igraph, or bn).
}
\examples{
plot.graph(obj, interactive=FALSE)
plot.graph(obj, interactive=TRUE)
}
\keyword{graph}
\keyword{plot}
