% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_feature_graph}
\alias{plot_feature_graph}
\title{Feature Graph Plotting}
\usage{
plot_feature_graph(
  x,
  genes,
  feature,
  from = NULL,
  feature.color = "red",
  other.color = "green",
  layout = NULL,
  interactive = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{genes}{Geneset with features.}

\item{from}{Input format (optional).}

\item{feature.color}{Color of the nodes with the feature (optional).}

\item{other.color}{Color of the other nodes (optional).}

\item{layout}{igraph plot layout (optional). It will be ignored if interactive=TRUE.}

\item{interactive}{Interactive plot (optional). Default: FALSE}

\item{features}{Feature whose graph you want to plot.}
}
\description{
This function allows you to plot the graph of a feature (e.g. transcription factors or tumor suppressors).
}
\examples{
plot_graph(obj, genes, 'tf', interactive=FALSE)
plot_graph(obj, genes, 'tumor.suppressor', interactive=TRUE)
}
\keyword{feature}
\keyword{graph}
\keyword{plot}
