% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_graph}
\alias{plot_graph}
\title{Graph Plotting}
\usage{
plot_graph(x, from = NULL, layout = NULL, interactive = FALSE)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Interactive plot (optional). Default: FALSE}

\item{layout}{igraph plot layout (optional). It will be ignored if interactive=TRUE.}
}
\description{
This function allows you to plot a graph, regardless of the format (adjacency matrix, list of edges, igraph, or bn).
}
\examples{
plot_graph(obj, interactive=FALSE)
plot_graph(obj, interactive=TRUE)
}
\keyword{graph}
\keyword{plot}
