% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{random.graph}
\alias{random.graph}
\title{Generate A Random Graph/DAG}
\usage{
random.graph(
  nodes,
  exp.degree,
  dag = TRUE,
  plot = TRUE,
  algorithm = c("regular", "watts", "er", "power", "bipartite", "barabasi",
    "geometric"),
  to = "igraph",
  ...
)
}
\arguments{
\item{nodes}{Number of nodes or vector of node names.}

\item{exp.degree}{Number of expected degree per node.}

\item{dag}{Whether the graph should be a DAG or not. Default: TRUE}

\item{plot}{Whether or not to plot the graph. Default: TRUE}

\item{algorithm}{Algorithm to be used to generate the graph: 'regular', 'watts', 'er', 'power', 'bipartite', 'barabasi', or 'geometric'. Default: 'regular'}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn').}
}
\description{
This function allows you to generate a random graph/DAG.
}
\examples{
graph <- random.graph(LETTERS[1:15], 3, dag=TRUE)
graph <- random.graph(LETTERS[1:15], 3, dag=FALSE)
graph <- random.graph(15, 3, dag=TRUE)
graph <- random.graph(15, 3, dag=FALSE)
}
\keyword{generate}
\keyword{graph}
\keyword{random}
