% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rename.genes}
\alias{rename.genes}
\title{Rename Genes To A Graph}
\usage{
rename.genes(
  g,
  genes,
  new.names,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{genes}{Genes to be renamed (vector).}

\item{new.names}{New names (vector, in the same order).}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function renames genes of a previously learned graph.
}
\examples{
g <- rename.genes(g, c('A', 'B', 'C'), c('Cd74', 'Ldhc', 'Tlx3'))
}
\keyword{genes}
\keyword{graph}
