% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rename.graphs}
\alias{rename.graphs}
\title{Rename Vertices Of A List Of Graphs}
\usage{
rename.graphs(graphs, names, to = "igraph")
}
\arguments{
\item{graphs}{List of graphs.}

\item{names}{Vector with the new vertex names (in order).}

\item{to}{Output format (('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn')) (optional).}
}
\description{
This function allows you to rename nodes of a list of graphs.
}
\examples{
graph <- rename.graphs(graphs, names)
}
\keyword{graph}
\keyword{rename}
