% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{run.glasso}
\alias{run.glasso}
\title{Run GLASSO Algorithm}
\usage{
run.glasso(
  df,
  rho = 0.5,
  to = "igraph",
  upper = FALSE,
  lower = TRUE,
  loops = FALSE,
  non.zero.coeff = NULL,
  unconnected.nodes = FALSE
)
}
\arguments{
\item{df}{Dataset.}

\item{rho}{Non-negative regularization parameter for GLASSO.}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn') (optional).}

\item{upper}{Whether or not to ignore the upper triangular adjacency matrix (optional).}

\item{lower}{Whether or not to ignore the lower triangular adjacency matrix (optional).}

\item{loops}{Whether or not to ignore the diagonal of the adjacency matrix (optional).}

\item{non.zero.coeff}{If other than NULL, non-zero adjacency coefficients shall be replaced by the value provided (optional). Default: NULL}

\item{unconnected.nodes}{Include unconnected nodes (optional). Default: FALSE}

\item{from}{Input format (optional).}
}
\description{
This function allows you to learn an undirected graph from a dataset using the GLASSO algorithm.
}
\examples{
graph <- run.glasso(df)
}
\keyword{graph}
\keyword{learning}
