% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{run.skeleton}
\alias{run.skeleton}
\title{Run Peter & Clark Skeleton Algorithm}
\usage{
run.skeleton(
  df,
  whitelist = NULL,
  blacklist = NULL,
  alpha = 0.01,
  max.sx = Inf,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = "igraph",
  cluster = 4,
  implementation = c("pcalg", "bnlearn"),
  pcalg.indep.test = pcalg::gaussCItest,
  pcalg.u2pd = c("relaxed", "rand", "retry"),
  pcalg.conservative = FALSE,
  pcalg.maj.rule = FALSE,
  pcalg.solve.confl = FALSE,
  bnlearn.test = ci.tests,
  bnlearn.B = NULL
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{alpha}{Target nominal type I error rate. Default: 0.01}

\item{max.sx}{Maximum allowed size of the conditioning sets.}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'igraph', or 'bn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}

\item{implementation}{Peter & Clark algorithm implementation: 'pcalg' or 'bnlearn'. Default: 'pcalg'}

\item{pcalg.indep.test}{Conditional independence test to be used (pcalg implementation). Default: pcalg::gaussCItest}

\item{bnlearn.test}{Conditional independence test to be used (bnlearn implementation): 'cor', 'mc-cor', 'smc-cor', 'zf', 'mc-zf', 'smc-zf', 'mi-g', 'mc-mi-g', 'smc-mi-g', or 'mi-g-sh'. Default: 'cor'}

\item{bnlearn.B}{Number of permutations considered for each permutation test (bnlearn implementation).}
}
\description{
This function allows you to learn a directed graph from a dataset using the Peter & Clark skeleton algorithm (stable version).
}
\examples{
graph <- run.skeleton(df, implementation='pcalg')
graph <- run.skeleton(df, implementation='bnlearn')
}
\keyword{graph}
\keyword{learning}
