% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{score.graph}
\alias{score.graph}
\title{Compute The Score Of A Graph}
\usage{
score.graph(
  g,
  test.df,
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{test.df}{Test dataset.}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function allows you to compute the score of a previously learned graph.
}
\examples{
score.graph(g, test.df)
}
\keyword{graph}
\keyword{score}
