% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{select.cells}
\alias{select.cells}
\title{Select (automatically) the most appropriate cell rows of your dataset.}
\usage{
select.cells(
  df,
  max.cells = 2000,
  selection.criteria = c("non.zeros", "mean", "variance", "random")
)
}
\arguments{
\item{df}{Dataset.}

\item{max.cells}{Maximum number of selected cells (optional).}

\item{selection.criteria}{Criteria to be applied to select cells. Default: 'non.zeros'}
}
\description{
This function allows you to (automatically) select the most appropriate cell rows of your dataset.
}
\examples{
df <- select.cells(df, max.cells=2000)
}
\keyword{cells}
\keyword{select}
