% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{select.genes}
\alias{select.genes}
\title{Select (automatically) the most appropriate gene columns of your dataset.}
\usage{
select.genes(
  df,
  genes = NULL,
  selected.genes = NULL,
  features = NULL,
  max.genes = 100,
  selection.criteria = c("non.zeros", "mean", "variance", "random"),
  min.non.zeros = 10,
  cor = TRUE,
  cor.method = c("spearman", "kendall", "pearson"),
  cor.threshold = 0.25
)
}
\arguments{
\item{df}{Dataset.}

\item{genes}{Geneset with features (optional).}

\item{selected.genes}{User-selected genes (optional).}

\item{features}{Features you want to select (optional).}

\item{max.genes}{Maximum number of selected genes (optional).}

\item{selection.criteria}{Criteria to be applied when it is necessary to select a maximum number of genes from the previously filtered ones. Default: 'non.zeros'}

\item{min.non.zeros}{Minimum number of non-zero values per gene (optional).}

\item{cor}{Using correlation to exclude unconnected genes (optional).}

\item{cor.method}{Correlation method: 'spearman', 'kendall', or 'pearson'. Default: 'spearman'}

\item{cor.threshold}{Correlation threshold. Default: 0.1}
}
\description{
This function allows you to (automatically) select the most appropriate gene columns of your dataset.
}
\examples{
df <- select.genes(df, max.genes=100, min.non.zeros=2)
df <- select.genes(df, genes=genes, features=c('tumor.suppressor', 'breast.cancer'), cor=TRUE, cor.threshold=0.7)
}
\keyword{genes}
\keyword{select}
