% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{skeleton}
\alias{skeleton}
\title{Peter & Clark Skeleton Algorithm}
\usage{
skeleton(
  df,
  whitelist = NULL,
  blacklist = NULL,
  alpha = 0.01,
  max.sx = Inf,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = parallel::detectCores(),
  implementation = c("pcalg", "bnlearn"),
  pcalg.indep.test = pcalg::gaussCItest,
  pcalg.u2pd = c("relaxed", "rand", "retry"),
  pcalg.conservative = FALSE,
  pcalg.maj.rule = FALSE,
  pcalg.solve.confl = FALSE,
  bnlearn.test = ci.tests,
  bnlearn.B = NULL,
  seed = sample(1:10^6, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{alpha}{Target nominal type I error rate. Default: 0.01}

\item{max.sx}{Maximum allowed size of the conditioning sets.}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: parallel::detectCores()}

\item{implementation}{Peter & Clark algorithm implementation: 'pcalg' or 'bnlearn'. Default: 'pcalg'}

\item{pcalg.indep.test}{Conditional independence test to be used (pcalg implementation). Default: pcalg::gaussCItest}

\item{bnlearn.test}{Conditional independence test to be used (bnlearn implementation): 'cor', 'mc-cor', 'smc-cor', 'zf', 'mc-zf', 'smc-zf', 'mi-g', 'mc-mi-g', 'smc-mi-g', or 'mi-g-sh'. Default: 'cor'}

\item{bnlearn.B}{Number of permutations considered for each permutation test (bnlearn implementation).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Peter & Clark skeleton algorithm (stable version).
}
\examples{
g <- skeleton(df, implementation='pcalg')
g <- skeleton(df, implementation='bnlearn')
}
\keyword{graph}
\keyword{learning}
