% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{split.dataframe}
\alias{split.dataframe}
\title{Split A Dataframe Into Training And Test Datasets}
\usage{
\method{split}{dataframe}(df, m = NULL)
}
\arguments{
\item{df}{Dataset.}

\item{m}{Size of training dataset (optional). Default: nrow(df)/2}
}
\description{
This function allows you to split a dataframe into training and test datasets.
}
\examples{
splitted.df <- valid.split(df)
splitted.df$train
splitted.df$test
}
\keyword{dataframe}
\keyword{test}
\keyword{training}
