% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{tabu}
\alias{tabu}
\title{Tabu Search Algorithm (TABU)}
\usage{
tabu(
  df,
  start = NULL,
  whitelist = NULL,
  blacklist = NULL,
  score = scores,
  tabu = 10,
  max.tabu = NULL,
  max.iter = Inf,
  maxp = Inf,
  m = NULL,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  seed = sample(1:10^10, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{start}{Preseeded directed acyclic graph used to initialize the algorithm (optional).}

\item{whitelist}{A data frame with two columns, containing a set of arcs to be included in the graph (optional).}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{score}{Score to be used: 'pred-loglik-g', 'loglik-g', 'aic-g', 'bic-g', or 'bge'. Default: 'pred-loglik-g'}

\item{tabu}{Length of the tabu list. Default: 10}

\item{max.tabu}{Iterations tabu search can perform without improving the best score. Default: tabu (10)}

\item{max.iter}{Maximum number of iterations. Default: Inf}

\item{maxp}{Maximum number of parents for a node. Default: Inf}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Tabu Search algorithm.
}
\examples{
g <- tabu(df)
}
\keyword{graph}
\keyword{learning}
