% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{test.function}
\alias{test.function}
\title{A Test Function}
\usage{
test.function(message = "hello world!")
}
\arguments{
\item{message}{Say something.}
}
\description{
This function allows you to test your package.
}
\examples{
test.function()
test.function('The way to get started is to quit talking and begin doing.')
test.function(message='Your time is limited, so don't waste it living someone else's life.')
}
\keyword{message}
\keyword{test}
