% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{undirected.edges}
\alias{undirected.edges}
\title{Return Undirected Edges}
\usage{
undirected.edges(
  g,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function returns all undirected edges in a graph.
}
\examples{
g <- undirected.edges(g)
}
\keyword{edges}
\keyword{graph}
\keyword{undirected}
