% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{wb.edgelist}
\alias{wb.edgelist}
\title{Select (automatically) the most appropriate gene columns of your dataset.}
\usage{
wb.edgelist(genes, from.features, to.features, selected.genes = NULL)
}
\arguments{
\item{genes}{Geneset with features.}

\item{from.features}{The edges will go from genes with some of these features.}

\item{to.features}{The edges will go to genes with some of these features.}

\item{selected.genes}{User-selected genes (optional).}
}
\description{
This function allows you to (automatically) select the most appropriate gene columns of your dataset.
}
\examples{
blacklist <- wb.edgelist(genes, from.features='target', to.features='tf')
}
\keyword{blacklist}
\keyword{genes}
\keyword{whitelist}
