% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.bnlearn}
\alias{as.bnlearn}
\title{Convert Graph To bnlearn Format}
\usage{
as.bnlearn(
  g,
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{from}{Input format (optional): 'auto', 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'auto'}
}
\description{
This function allows you to convert your graph to bnlearn format.
}
\examples{
g <- as.bnlearn(g)
}
