% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.edges}
\alias{as.edges}
\title{Convert Graph To Edge List}
\usage{
as.edges(g, from = c(NULL, "adjacency", "edges", "graph", "igraph", "bnlearn"))
}
\arguments{
\item{g}{Graph object.}

\item{from}{Input format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: autodetect}
}
\description{
This function allows you to convert your graph to edge list format.
}
\examples{
g <- as.edges(g)
}
