% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as.igraph}
\alias{as.igraph}
\title{Convert Graph To igraph Format}
\usage{
as.igraph(
  g,
  from = c(NULL, "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{from}{Input format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: autodetect}
}
\description{
This function allows you to convert your graph to igraph format.
}
\examples{
g <- as.igraph(g)
}
