% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{averaged.graph}
\alias{averaged.graph}
\title{Calculate The Averaged Graph}
\usage{
averaged.graph(graphs, threshold = 0.5, to = "igraph")
}
\arguments{
\item{graphs}{List of graphs.}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged graph (optional). Default: 0.5}

\item{to}{Output format (('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn')) (optional).}
}
\description{
This function allows you to calculate the averaged graph from a list of graphs.
}
\examples{
graph <- averaged.graph(graphs)
}
\keyword{average}
\keyword{graph}
