% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.ges}
\alias{boot.ges}
\title{Greedy Equivalence Search Algorithm (GES) With Bootstrapping}
\usage{
boot.ges(
  df,
  blacklist = NULL,
  adaptive = c("none", "vstructures", "triples"),
  maxDegree = integer(0),
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = parallel::detectCores(),
  seed = sample(1:10^6, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{blacklist}{A data frame with two columns, containing a set of arcs not to be included in the graph (optional).}

\item{adaptive}{Whether constraints should be adapted to newly detected v-structures or unshielded triples: 'none', 'vstructures', or 'triples'. Default: 'none'}

\item{maxDegree}{Parameter used to limit the vertex degree of the estimated graph. Default: integer(0)}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: parallel::detectCores()}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn a directed graph from a dataset using the Greedy Equivalence Search (GES) algorithm of Chickering (2002).
}
\examples{
obj <- boot.ges(df)
avg.g <- obj$average
g.rep <- obj$replicates
}
\keyword{graph}
\keyword{learning}
