% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.glasso}
\alias{boot.glasso}
\title{Graphical Lasso (GLASSO) With Bootstrapping}
\usage{
boot.glasso(
  df,
  rho = 0.1,
  R = 200,
  m = NULL,
  threshold = 0.5,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = parallel::detectCores(),
  seed = sample(1:10^6, 1)
)
}
\arguments{
\item{df}{Dataset.}

\item{rho}{Non-negative regularization parameter for GLASSO.}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of training set (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the average adjacency matrix (optional). Default: 0.5}

\item{to}{Output format ('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn') (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: parallel::detectCores()}

\item{seed}{Seed used for random selection. Default: NULL}
}
\description{
This function allows you to learn an undirected graph from a dataset using the GLASSO algorithm.
}
\examples{
obj <- boot.glasso(df, rho=0.1)
avg.g <- obj$average
g.rep <- obj$replicates
}
\keyword{graph}
\keyword{learning}
