% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{boot.glasso}
\alias{boot.glasso}
\title{GLASSO Algorithm With Bootstrapping}
\usage{
boot.glasso(
  df,
  rho = 0.1,
  R = 200,
  m = NULL,
  threshold = 0.5,
  upper = FALSE,
  lower = TRUE,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  cluster = 4
)
}
\arguments{
\item{df}{Dataset.}

\item{rho}{Non-negative regularization parameter for GLASSO.}

\item{R}{Number of bootstrap replicates (optional). Default: 200}

\item{m}{Size of each bootstrap replicate (optional). Default: nrow(df)/2}

\item{threshold}{Minimum strength required for a coefficient to be included in the averaged adjacency matrix (optional). Default: 0.5}

\item{upper}{Whether or not to ignore the upper triangular adjacency matrix (optional).}

\item{lower}{Whether or not to ignore the lower triangular adjacency matrix (optional).}

\item{to}{Output format (('adjacency', 'edges', 'graph', 'igraph', or 'bnlearn')) (optional).}

\item{cluster}{A cluster object from package parallel or the number of cores to be used (optional). Default: 4}
}
\description{
This function allows you to learn an undirected graph from a dataset using the GLASSO algorithm.
}
\examples{
graph <- boot.glasso(df, rho=0.1)
}
\keyword{graph}
\keyword{learning}
