% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compare.graphs}
\alias{compare.graphs}
\title{Graph Comparison}
\usage{
compare.graphs(
  learned,
  true,
  learned.replicates = NULL,
  skeleton = FALSE,
  arcs = FALSE,
  plot = TRUE,
  vertical.plot = TRUE,
  split.plot = TRUE
)
}
\arguments{
\item{learned}{Learned graph or graph 1 (obj$average).}

\item{true}{Ground truth graph or graph 2 (reference).}

\item{learned.replicates}{List of learned replicates (obj$replicates) (optional). If provided, PR AUC will be calculated.}

\item{skeleton}{Whether to compare graph skeletons instead of the graphs themself. Default: FALSE}

\item{arcs}{Whether or not to list the arcs. Default: FALSE.}

\item{plot}{Whether or not to plot the differences between the two graphs. Default: TRUE}

\item{vertical.plot}{Whether to draw the comparison plots horizontally. Otherwise, they will be drawn horizontally. Default: TRUE}

\item{split.plot}{Whether to split comparison plots. Otherwise, they will be drawn together. Default: TRUE}
}
\description{
This function allows you to compare two graphs.
}
\examples{
comparison <- compare.graphs(obj1, obj2, plot=TRUE)
comparison <- compare.graphs(obj1, obj2, plot=FALSE)
}
\keyword{comparison}
\keyword{graph}
