% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compare.graphs}
\alias{compare.graphs}
\title{Graph Comparison}
\usage{
compare.graphs(learned, true, arcs = FALSE)
}
\arguments{
\item{learned}{Learned graph or graph 1.}

\item{true}{Ground truth graph or graph 2 (reference).}

\item{arcs}{Whether or not to list the arcs (optional). Default: FALSE.}
}
\description{
This function allows you to compare two graphs, regardless of the format (adjacency matrix, list of edges, igraph, or bn).
}
\examples{
comparison <- compare.graphs(obj1, obj2)
}
\keyword{comparison}
\keyword{graph}
