% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert.format}
\alias{convert.format}
\title{Graph Format Conversion}
\usage{
convert.format(
  g,
  to = c("igraph", "adjacency", "edges", "graph", "bnlearn"),
  from = c(NULL, "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{g}{Graph object.}

\item{to}{Output format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: 'igraph'}

\item{from}{Input format (optional): 'adjacency', 'edges', 'graph', 'igraph', or 'bnlearn'. Default: autodetect}
}
\description{
This function allows you to convert graph between different formats: adjacency matrix, list of edges, igraph and bnlearn (bnlearn).
}
\examples{
g <- convert.format(g, to='igraph')
}
\keyword{adjacency}
\keyword{edges}
\keyword{format}
\keyword{graph}
