% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.communities}
\alias{graph.communities}
\title{Graph Communities}
\usage{
graph.communities(x, from = NULL)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}
}
\description{
This function allows you to detect how many communities are in the graph and to which community each node and edge belongs.
}
\examples{
communities <- graph.communities(g)
}
\keyword{community}
\keyword{graph}
