% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.communities}
\alias{graph.communities}
\title{Graph Communities}
\usage{
graph.communities(
  x,
  algorithm = c("louvain", "edge.betweenness", "fast.greedy", "label.prop",
    "leading.eigen", "optimal", "spinglass", "walktrap"),
  dendrogram = TRUE,
  from = NULL
)
}
\arguments{
\item{x}{Graph object.}

\item{algorithm}{Algorithm for finding communities: 'louvain', 'edge.betweenness', 'fast.greedy', 'label.prop', 'leading.eigen', 'optimal', 'spinglass', or 'walktrap'. Default: 'louvain'}

\item{dendrogram}{Whether or not to plot a dendrogram (when possible). Default: TRUE}

\item{from}{Input format (optional).}
}
\description{
This function allows you to detect how many communities are in the graph and to which community each node and edge belongs.
}
\examples{
communities <- graph.communities(g)
}
\keyword{community}
\keyword{graph}
