% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.communities}
\alias{graph.communities}
\title{Graph Communities}
\usage{
graph.communities(
  x,
  algorithm = c("louvain", "edge.betweenness", "fast.greedy", "label.prop",
    "leading.eigen", "optimal", "spinglass", "walktrap"),
  network = TRUE,
  network.layout = c("grid", "star", "circle", "sphere", "nicely"),
  interactive.network = FALSE,
  network.isolated = TRUE,
  dendrogram = FALSE,
  dendrogram.type = c("fan", "phylogram", "cladogram", "unrooted", "radial"),
  from = c("auto", "adjacency", "edges", "graph", "igraph", "bnlearn")
)
}
\arguments{
\item{x}{Graph object.}

\item{algorithm}{Algorithm for finding communities: 'louvain', 'edge.betweenness', 'fast.greedy', 'label.prop', 'leading.eigen', 'optimal', 'spinglass', or 'walktrap'. Default: 'louvain'}

\item{network}{Whether or not to plot the network. Default: TRUE}

\item{network.layout}{igraph network layout (optional): 'grid', 'star', 'circle', 'sphere', or 'nicely'. Default: 'grid'}

\item{interactive.network}{Interactive network (optional). Default: FALSE}

\item{network.isolated}{Whether or not to include isolated nodes in the plot (optional). Default: TRUE}

\item{dendrogram}{Whether or not to plot a dendrogram (when possible). Default: FALSE}

\item{dendrogram.type}{Type of phylogeny to be drawn: 'fan', 'phylogram', 'cladogram', 'unrooted', or 'radial'. Default: 'fan'}

\item{from}{Input format (optional).}
}
\description{
This function allows you to detect how many communities are in the graph and to which community each node and edge belongs.
}
\examples{
communities <- graph.communities(g)
communities <- graph.communities(g, algorithm='louvain', network=TRUE, network.isolated=FALSE, dendrogram=FALSE)
communities <- graph.communities(g, algorithm='walktrap', network=FALSE, dendrogram=TRUE, dendrogram.type='cladogram')
}
\keyword{community}
\keyword{graph}
\keyword{plot}
