% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graph.plot}
\alias{graph.plot}
\title{Graph Plotting}
\usage{
graph.plot(
  x,
  from = NULL,
  layout = c("grid", "star", "circle", "tree", "nicely"),
  interactive = FALSE
)
}
\arguments{
\item{x}{Graph object.}

\item{from}{Input format (optional).}

\item{layout}{igraph plot layout (optional): 'grid', 'star', 'circle', 'tree', or 'nicely'. It will be ignored if interactive=TRUE. Default: 'grid'}

\item{interactive}{Interactive plot (optional). Default: FALSE}
}
\description{
This function allows you to plot a graph, regardless of the format (adjacency matrix, list of edges, igraph, or bnlearn).
}
\examples{
graph.plot(obj, interactive=FALSE)
graph.plot(obj, interactive=TRUE)
}
\keyword{graph}
\keyword{plot}
